`double` was `f32` for eighty-eight rounds; Dart's is `f64`
The user found it. Dart's `double` is IEEE-754 double precision -- the
language says so -- and this compiler mapped it to f32 from its first
round. The old note gave the reason: the hand port measures in f32
because that is what the engine's geometry takes, so a translated value
type has to sit beside one. That is a fact about the hand port, and it
was allowed to decide what a `double` *is*, which is not a translator's
choice to make. Where a translated value meets the engine's f32, the
cast belongs at that boundary.

Nothing caught it, and the reason is worth keeping. Every fixture value
is exactly representable in both widths, so 130 tests passed. Both front
ends share the backend, so they agreed -- wrongly, together. And f32 is
a perfectly legal type, so the output compiled. All three instruments
are blind to this, which makes it the first entry here where everything
passed and the answer was still wrong.

Changing it moved the package's error count by zero, which is the proof
that it was silent. `int` checked at the same time: i64 is right, since
Dart's int on the VM and AOT targets is a 64-bit two's complement
integer.

Also this round: dart:collection's internal classes as *types* and not
only as constructors; unnamed factories, which arrive as static calls
with an empty name and had been walking into the operator table to say
"operator `` has no Rust name" 367 times; and int.parse, double.parse
and the collection constructors reached as statics.

Errors 675 -> 814, refusals 5399 -> 5179.

+++ tools/dart2rust/lib/prelude.dart
@@ -797,13 +797,13 @@ impl ByteData {
         self.bytes[at..at + 4].copy_from_slice(&(value as u32).to_le_bytes());
     }
 
-    pub fn get_float32(&self, at: i64) -> f32 {
-        f32::from_le_bytes(self.four(at, 4)[..4].try_into().unwrap())
+    pub fn get_float32(&self, at: i64) -> f64 {
+        f32::from_le_bytes(self.four(at, 4)[..4].try_into().unwrap()) as f64
     }
 
-    pub fn set_float32(&mut self, at: i64, value: f32) {
+    pub fn set_float32(&mut self, at: i64, value: f64) {
         let at = at as usize;
-        self.bytes[at..at + 4].copy_from_slice(&value.to_le_bytes());
+        self.bytes[at..at + 4].copy_from_slice(&(value as f32).to_le_bytes());
     }
 
     pub fn get_float64(&self, at: i64) -> f64 {
